Small changes in response to comments in the PR
