+++ packages/lib/src/components/Ach/types.ts
@@ -1,5 +1,4 @@
 import { UIElementProps } from '../internal/UIElement/types';
-import { ValidationRuleResult } from '../../utils/Validator/ValidationRuleResult';
 
 export interface AchConfiguration extends UIElementProps {
     /**
@@ -34,20 +33,4 @@ export interface AchPlaceholders {
     routingNumber?: string;
     accountNumber?: string;
     accountNumberVerification?: string;
-}
-
-export interface AchStateErrors {
-    selectedAccountType: ValidationRuleResult;
-    ownerName: ValidationRuleResult;
-    routingNumber: ValidationRuleResult;
-    accountNumber: ValidationRuleResult;
-    accountNumberVerification: ValidationRuleResult;
-}
-
-export interface AchStateData {
-    selectedAccountType: string;
-    ownerName: string;
-    routingNumber: string;
-    accountNumber: string;
-    accountNumberVerification: string;
 }
